% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sql.R
\name{get_sql}
\alias{get_sql}
\title{Get underlying sql for a Looker query.}
\usage{
get_sql(model, view, fields, filters = list(), limit = 1000)
}
\arguments{
\item{model}{character. The \code{model} parameter of the query.}

\item{view}{character. The \code{view} parameter of the query.}

\item{fields}{character. The \code{fields} parameter of the query.}

\item{filters}{list or character.  Either a named list, or a character
vector using colons as separators
each vector describing one of the \code{filters} of the query.}

\item{limit}{numeric. The \code{limit} parameter of the query.}
}
\value{
a simple string containing the SQL query.
}
\description{
This function assumes that you have set the environment
variables \code{LOOKER_URL}, \code{LOOKER_ID}, and
\code{LOOKER_SECRET}.
}
\details{
The function will not allow the query to proceed unless you specify
a model, a view, and at least one field.
}
\examples{
\dontrun{
  df <- get_sql(model = "thelook",
                view = "orders",
                fields = c("orders.count", "orders.created_month")
                filters = list("orders.created_month" = "90 days", "orders.status" = "complete")
  )

  df <- get_sql(model = "thelook",
                view = "orders",
                fields = c("orders.count", "orders.created_month")
                filters = c("orders.created_month: 90 days", "orders.status: complete")
  )
}
}

