% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looker3-package.R, R/looker3.R
\docType{package}
\name{looker3}
\alias{looker3}
\alias{looker3-package}
\title{Describe the package.}
\usage{
looker3(model, view, fields, filters = list(), limit = 1000)
}
\arguments{
\item{model}{character. The \code{model} parameter of the query.}

\item{view}{character. The \code{view} parameter of the query.}

\item{fields}{character. The \code{fields} parameter of the query.}

\item{filters}{list. Each element of the list is a length 2 character vector,
each vector describing one of the \code{filters} of the query.}

\item{limit}{numeric. The \code{limit} parameter of the query.}
}
\value{
a data.frame containing the data returned by the query
}
\description{
Describe the package.

This function assumes that you have set the environment
variables \code{LOOKER_URL}, \code{LOOKER_ID}, and 
\code{LOOKER_SECRET}.
}
\details{
The function will not allow the query to proceed unless you specify
a model, a view, and at least one field.
}

