% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inline_query.R
\name{run_inline_query}
\alias{run_inline_query}
\title{Run an inline query using the Looker 3.0 API.}
\usage{
run_inline_query(base_url, client_id, client_secret, model, view, fields,
  filters, limit = 1000)
}
\arguments{
\item{base_url}{character. The base url of the Looker server.}

\item{client_id}{character. Your client id.}

\item{client_secret}{character. Your client secret.}

\item{model}{character. The \code{model} parameter of the query.}

\item{view}{character. The \code{view} parameter of the query.}

\item{fields}{character. The \code{fields} parameter of the query.}

\item{filters}{named list. The names and entries of the list will
be converted into a hash table and passed as the \code{filters}
of the query.}

\item{limit}{numeric. The \code{limit} parameter of the query.}
}
\value{
a data.frame containing the data returned by the query.
}
\description{
Run an inline query using the Looker 3.0 API.
}

